requirements.txt
flask
ollama
chromadb
sentence-transformers
pypdf
pytesseract
pillow
watchdog
psutil
